# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FLOW API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/investigations": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "List Findings",
        "description": "列出全部 Finding（含身份交接标识），供「分析与归因」入口选择调查对象。",
        "operationId": "list_findings_api_v1_investigations_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{finding_id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Investigation Context",
        "operationId": "investigation_context_api_v1_investigations__finding_id__get",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          },
          {
            "name": "batch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Batch Id"
            }
          },
          {
            "name": "metric_snapshot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric Snapshot Id"
            }
          },
          {
            "name": "analysis_run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Analysis Run Id"
            }
          },
          {
            "name": "source_record_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 12,
              "title": "Source Record Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationContextResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{finding_id}/evidence/{evidence_id}/decision": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Decide Evidence",
        "operationId": "decide_evidence_api_v1_investigations__finding_id__evidence__evidence_id__decision_post",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          },
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDecisionResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{finding_id}/conclusion": {
      "put": {
        "tags": [
          "investigations"
        ],
        "summary": "Save Conclusion",
        "operationId": "save_conclusion_api_v1_investigations__finding_id__conclusion_put",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConclusionUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConclusionResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{finding_id}/transition": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Transition Finding",
        "operationId": "transition_finding_api_v1_investigations__finding_id__transition_post",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingTransitionResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConclusionResponse": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_sequence": {
            "type": "integer",
            "title": "Review Sequence"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "finding_id",
          "status",
          "review_sequence",
          "decision"
        ],
        "title": "ConclusionResponse",
        "description": "Acknowledgement of a stored Finance BP conclusion."
      },
...
      "ConclusionUpsertRequest": {
        "properties": {
          "verified_facts": {
            "type": "string",
            "minLength": 1,
            "title": "Verified Facts"
          },
          "analysis_judgment": {
            "type": "string",
            "minLength": 1,
            "title": "Analysis Judgment"
          },
          "open_questions": {
            "type": "string",
            "minLength": 1,
            "title": "Open Questions"
          },
          "recommendation": {
            "type": "string",
            "minLength": 1,
            "title": "Recommendation"
          },
          "editor": {
            "type": "string",
            "minLength": 1,
            "title": "Editor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "verified_facts",
          "analysis_judgment",
          "open_questions",
          "recommendation",
          "editor"
        ],
        "title": "ConclusionUpsertRequest"
      },
...
      "DriverLine": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "driver_code": {
            "type": "string",
            "title": "Driver Code"
          },
          "calculation_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calculation Method"
          },
          "contribution_amount": {
            "type": "string",
            "title": "Contribution Amount"
          },
          "contribution_ratio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution Ratio"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "position",
          "driver_code",
          "calculation_method",
          "contribution_amount",
          "contribution_ratio"
        ],
        "title": "DriverLine"
      },
...
      "EvidenceDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "reviewer": {
            "type": "string",
            "minLength": 1,
            "title": "Reviewer"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "reviewer"
        ],
        "title": "EvidenceDecisionRequest"
      },
      "EvidenceDecisionResponse": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_sequence": {
            "type": "integer",
            "title": "Review Sequence"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "finding_id",
          "status",
          "review_sequence",
          "decision"
        ],
        "title": "EvidenceDecisionResponse",
        "description": "Acknowledgement of an evidence review decision."
      },
...
      "FindingContext": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "finding_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finding Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "impact_amount": {
            "type": "string",
            "title": "Impact Amount"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "business_meaning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Meaning"
          },
          "fact_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Statement"
          },
          "comparison_basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparison Basis"
          },
          "total_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Score"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "finding_id",
          "finding_type",
          "title",
          "status",
          "impact_amount",
          "unit",
          "confidence",
          "business_meaning",
          "fact_statement",
          "comparison_basis",
          "total_score",
          "policy_version",
          "created_at"
        ],
        "title": "FindingContext"
      },
...
      "FindingListItem": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "finding_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finding Type"
          },
          "impact_amount": {
            "type": "string",
            "title": "Impact Amount"
          },
          "comparison_basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparison Basis"
          },
          "total_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Score"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "metric_snapshot_id": {
            "type": "string",
            "title": "Metric Snapshot Id"
          },
          "analysis_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Run Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "finding_id",
          "title",
          "status",
          "finding_type",
          "impact_amount",
          "comparison_basis",
          "total_score",
          "batch_id",
          "metric_snapshot_id",
          "analysis_run_id",
          "created_at"
        ],
        "title": "FindingListItem",
        "description": "调查列表项：携带 Investigation 身份交接所需的全部标识（D036）。"
      },
      "FindingListResponse": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingListItem"
            },
            "type": "array",
            "title": "Findings"
          }
        },
        "type": "object",
        "required": [
          "findings"
        ],
        "title": "FindingListResponse"
      },
      "FindingTransitionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "reviewer": {
            "type": "string",
            "minLength": 1,
            "title": "Reviewer"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "reviewer"
        ],
        "title": "FindingTransitionRequest"
      },
      "FindingTransitionResponse": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_sequence": {
            "type": "integer",
            "title": "Review Sequence"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "finding_id",
          "status",
          "review_sequence",
          "decision"
        ],
        "title": "FindingTransitionResponse",
        "description": "Acknowledgement of a governed Finding state transition."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationContextResponse": {
        "properties": {
          "identity": {
            "$ref": "#/components/schemas/InvestigationIdentity"
          },
          "finding": {
            "$ref": "#/components/schemas/FindingContext"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisResultContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "metric": {
            "$ref": "#/components/schemas/MetricContext"
          },
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/DriverLine"
            },
            "type": "array",
            "title": "Drivers"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLine"
            },
            "type": "array",
            "title": "Evidence"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewLine"
            },
            "type": "array",
            "title": "Reviews"
          },
          "quality_issues": {
            "items": {
              "$ref": "#/components/schemas/QualityIssueLine"
            },
            "type": "array",
            "title": "Quality Issues"
          },
          "reconciliations": {
            "items": {
              "$ref": "#/components/schemas/ReconciliationLine"
            },
            "type": "array",
            "title": "Reconciliations"
          },
          "conclusion": {
            "$ref": "#/components/schemas/ConclusionState"
          },
          "source_records": {
            "items": {
              "$ref": "#/components/schemas/SourceRecordLine"
            },
            "type": "array",
            "title": "Source Records"
          },
          "eligibility_blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Eligibility Blockers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identity",
          "finding",
          "result",
          "metric",
          "drivers",
          "evidence",
          "reviews",
          "quality_issues",
          "reconciliations",
          "conclusion",
          "source_records",
          "eligibility_blockers"
        ],
        "title": "InvestigationContextResponse",
        "description": "Read-only evidence-first Investigation context."
      },
      "InvestigationErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "InvestigationErrorResponse"
      },
      "InvestigationIdentity": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "metric_snapshot_id": {
            "type": "string",
            "title": "Metric Snapshot Id"
          },
          "analysis_run_id": {
            "type": "string",
            "title": "Analysis Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "finding_id",
          "batch_id",
          "metric_snapshot_id",
          "analysis_run_id"
        ],
        "title": "InvestigationIdentity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}